Refactor author picker styles and structure; streamline SCSS files, enhance layout, and improve component integration
