+++ package.json
@@ -41,6 +41,7 @@
 		"@wordpress/blocks": "^14.11.0",
 		"@wordpress/components": "^29.8.0",
 		"@wordpress/compose": "^7.22.0",
+		"@wordpress/core-data": "^7.22.0",
 		"@wordpress/data": "^10.22.0",
 		"@wordpress/element": "^6.22.0",
 		"@wordpress/eslint-plugin": "^22.8.0",
