+++ src/blocks/author-profile/style.scss
@@ -7,6 +7,7 @@
 
 // Main block container
 .wp-block-author-profile-blocks-author-profile {
+  width: 100%;
   padding: c.$spacing-md;
 
   .apb-author-profile-content {
