--- src/components/index.js
@@ -1,4 +1,0 @@
-/**
- * Export components for reuse
- */
-export { AuthorPicker, AuthorBlockPlaceholder } from '../js/components';
