+++ src/scss/common/_placeholder.scss
@@ -3,7 +3,7 @@
  */
 
 .apb-author-block-placeholder {
-    padding: 24px;
+    //padding: 24px;
     text-align: center;
     background-color: #f8f9fa;
     border: 1px dashed #ddd;
