+++ src/scss/components/author-picker.scss
@@ -1,9 +1,6 @@
-/**
- * Shared author-picker component styles
- */
 
 // Import common styles
-@use '../scss/common' as c;
+@use '../../scss/common' as c;
 
 .apb-author-picker-wrapper {
